A slightly more succinct version of the code :)
I am doing the intermediate C++ course, coming from coding for about a year in Java. And I gave this task much thought and came up with this logic that makes the code fairly short and contained. Focusing on all the lose scenarios, and then everything else is a win...or a tie which is taken care of in the first if().  Let me know what you think.